Move non-public files to Internal folder and namespace
