+++ src/EnumScribe/Internal/EnumScribeConsts.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.CodeAnalysis;
 using System.Collections.Generic;
 
-namespace EnumScribe
+namespace EnumScribe.Internal
 {
     internal static class EnumScribeConsts
     {
