+++ src/EnumScribe/Internal/EnumScribeSyntaxReceiver.cs
@@ -3,7 +3,7 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Linq;
 using System.Collections.Generic;
 
-namespace EnumScribe
+namespace EnumScribe.Internal
 {
     internal sealed class EnumScribeSyntaxReceiver : ISyntaxContextReceiver
     {
