+++ src/EnumScribe/Internal/PublicAPI.Unshipped.txt
